ResultadoBusqueda, AsistenciaDa, VistaAsistencia,  Busqueda, ControladorAsistencia
resultadoBusqueda eliminado, agregación parcial método búsqueda
