--- src/root/gestionagricola/vistas/Busqueda.java
@@ -1,193 +1,0 @@
-
-package root.gestionagricola.vistas;
-
-/**
- *
- * @author Javier
- */
-public class Busqueda extends javax.swing.JPanel {
-    
-    
-
-    private ControladorVistas controladorVista;
-    private String panel_anterior;
-    public Busqueda() {
-        initComponents();
-
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        Titulo7 = new javax.swing.JLabel();
-        jTextField4 = new javax.swing.JTextField();
-        Titulo8 = new javax.swing.JLabel();
-        Titulo9 = new javax.swing.JLabel();
-        Titulo11 = new javax.swing.JLabel();
-        jButton4 = new javax.swing.JButton();
-        jTextField6 = new javax.swing.JTextField();
-        Titulo12 = new javax.swing.JLabel();
-        jButton6 = new javax.swing.JButton();
-        jLabel2 = new javax.swing.JLabel();
-        jTextField7 = new javax.swing.JTextField();
-        Titulo14 = new javax.swing.JLabel();
-        jTextField8 = new javax.swing.JTextField();
-        jButton1 = new javax.swing.JButton();
-        jDateChooser1 = new com.toedter.calendar.JDateChooser();
-        Fondo = new javax.swing.JLabel();
-
-        setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
-
-        Titulo7.setBackground(new java.awt.Color(255, 255, 255));
-        Titulo7.setFont(new java.awt.Font("Garamond", 1, 36)); // NOI18N
-        Titulo7.setForeground(new java.awt.Color(255, 255, 255));
-        Titulo7.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        Titulo7.setText("Rut:");
-        add(Titulo7, new org.netbeans.lib.awtextra.AbsoluteConstraints(230, 240, 170, 60));
-        add(jTextField4, new org.netbeans.lib.awtextra.AbsoluteConstraints(530, 250, 320, 40));
-
-        Titulo8.setBackground(new java.awt.Color(255, 255, 255));
-        Titulo8.setFont(new java.awt.Font("Garamond", 1, 36)); // NOI18N
-        Titulo8.setForeground(new java.awt.Color(255, 255, 255));
-        Titulo8.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        Titulo8.setText("Fecha de ");
-        add(Titulo8, new org.netbeans.lib.awtextra.AbsoluteConstraints(260, 290, 200, 60));
-
-        Titulo9.setBackground(new java.awt.Color(255, 255, 255));
-        Titulo9.setFont(new java.awt.Font("Garamond", 1, 36)); // NOI18N
-        Titulo9.setForeground(new java.awt.Color(255, 255, 255));
-        Titulo9.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        Titulo9.setText("Nacimiento:");
-        add(Titulo9, new org.netbeans.lib.awtextra.AbsoluteConstraints(260, 340, 240, 60));
-
-        Titulo11.setBackground(new java.awt.Color(255, 255, 255));
-        Titulo11.setFont(new java.awt.Font("Garamond", 1, 36)); // NOI18N
-        Titulo11.setForeground(new java.awt.Color(255, 255, 255));
-        Titulo11.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        Titulo11.setText("Nombre:");
-        add(Titulo11, new org.netbeans.lib.awtextra.AbsoluteConstraints(260, 180, 180, 60));
-
-        jButton4.setIcon(new javax.swing.ImageIcon(getClass().getResource("/root/gestionagricola/vistas/imagenes/Volver atras.png"))); // NOI18N
-        jButton4.setBorderPainted(false);
-        jButton4.setContentAreaFilled(false);
-        jButton4.setFocusPainted(false);
-        jButton4.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButton4ActionPerformed(evt);
-            }
-        });
-        add(jButton4, new org.netbeans.lib.awtextra.AbsoluteConstraints(1020, 30, 130, 70));
-        add(jTextField6, new org.netbeans.lib.awtextra.AbsoluteConstraints(530, 190, 320, 40));
-
-        Titulo12.setBackground(new java.awt.Color(255, 255, 255));
-        Titulo12.setFont(new java.awt.Font("Garamond", 1, 36)); // NOI18N
-        Titulo12.setForeground(new java.awt.Color(255, 255, 255));
-        Titulo12.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        Titulo12.setText("Cargo:");
-        add(Titulo12, new org.netbeans.lib.awtextra.AbsoluteConstraints(230, 410, 210, 60));
-
-        jButton6.setBackground(new java.awt.Color(255, 255, 255));
-        jButton6.setFont(new java.awt.Font("Garamond", 1, 36)); // NOI18N
-        jButton6.setForeground(new java.awt.Color(0, 0, 0));
-        jButton6.setIcon(new javax.swing.ImageIcon(getClass().getResource("/root/gestionagricola/vistas/imagenes/Busqueda 32px.png"))); // NOI18N
-        jButton6.setText("Búsqueda");
-        jButton6.setBorder(new javax.swing.border.LineBorder(new java.awt.Color(255, 153, 153), 2, true));
-        jButton6.setFocusPainted(false);
-        jButton6.setHorizontalTextPosition(javax.swing.SwingConstants.LEFT);
-        jButton6.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButton6ActionPerformed(evt);
-            }
-        });
-        add(jButton6, new org.netbeans.lib.awtextra.AbsoluteConstraints(130, 110, 1020, 50));
-
-        jLabel2.setIcon(new javax.swing.ImageIcon(getClass().getResource("/root/gestionagricola/vistas/imagenes/frutap.png"))); // NOI18N
-        jLabel2.setText("jLabel2");
-        add(jLabel2, new org.netbeans.lib.awtextra.AbsoluteConstraints(-90, 0, 229, 151));
-        add(jTextField7, new org.netbeans.lib.awtextra.AbsoluteConstraints(530, 420, 320, 40));
-
-        Titulo14.setBackground(new java.awt.Color(255, 255, 255));
-        Titulo14.setFont(new java.awt.Font("Garamond", 1, 36)); // NOI18N
-        Titulo14.setForeground(new java.awt.Color(255, 255, 255));
-        Titulo14.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        Titulo14.setText("N° Contrato:");
-        add(Titulo14, new org.netbeans.lib.awtextra.AbsoluteConstraints(250, 480, 270, 60));
-        add(jTextField8, new org.netbeans.lib.awtextra.AbsoluteConstraints(530, 490, 320, 40));
-
-        jButton1.setBackground(new java.awt.Color(255, 255, 255));
-        jButton1.setFont(new java.awt.Font("Garamond", 1, 24)); // NOI18N
-        jButton1.setForeground(new java.awt.Color(0, 0, 0));
-        jButton1.setText("Realizar Búsqueda");
-        jButton1.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButton1ActionPerformed(evt);
-            }
-        });
-        add(jButton1, new org.netbeans.lib.awtextra.AbsoluteConstraints(530, 570, 260, 60));
-
-        jDateChooser1.setBackground(new java.awt.Color(255, 255, 255));
-        jDateChooser1.setForeground(new java.awt.Color(255, 255, 255));
-        jDateChooser1.setFont(new java.awt.Font("Garamond", 0, 12)); // NOI18N
-        add(jDateChooser1, new org.netbeans.lib.awtextra.AbsoluteConstraints(530, 330, 320, 40));
-
-        Fondo.setIcon(new javax.swing.ImageIcon(getClass().getResource("/root/gestionagricola/vistas/imagenes/Fondo oscuro.png"))); // NOI18N
-        Fondo.setPreferredSize(new java.awt.Dimension(1299, 735));
-        add(Fondo, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 1270, 735));
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void jButton6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton6ActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_jButton6ActionPerformed
-
-    private void jButton4ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton4ActionPerformed
-       //pueden acceder desde supervisor o dueño
-       //volver al dueño
-       if(this.panel_anterior.equals("busquedadueño")){
-           this.controladorVista.SeleccionarPanel("dueño");
-       }else if(this.panel_anterior.equals("busquedasupervisor")){//volver al supervisor
-            this.controladorVista.SeleccionarPanel("supervisor");
-       }
-       
-       
-
-    }//GEN-LAST:event_jButton4ActionPerformed
-
-    private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
-        this.controladorVista.SeleccionarPanel("resultadobusqueda");
-    }//GEN-LAST:event_jButton1ActionPerformed
-
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JLabel Fondo;
-    private javax.swing.JLabel Titulo11;
-    private javax.swing.JLabel Titulo12;
-    private javax.swing.JLabel Titulo14;
-    private javax.swing.JLabel Titulo7;
-    private javax.swing.JLabel Titulo8;
-    private javax.swing.JLabel Titulo9;
-    private javax.swing.JButton jButton1;
-    private javax.swing.JButton jButton4;
-    private javax.swing.JButton jButton6;
-    private com.toedter.calendar.JDateChooser jDateChooser1;
-    private javax.swing.JLabel jLabel2;
-    private javax.swing.JTextField jTextField4;
-    private javax.swing.JTextField jTextField6;
-    private javax.swing.JTextField jTextField7;
-    private javax.swing.JTextField jTextField8;
-    // End of variables declaration//GEN-END:variables
-
-    public void setControladorVista(ControladorVistas controladorVista) {
-        this.controladorVista = controladorVista;
-    }
-
-    public void setPanel_anterior(String panel_anterior) {
-        this.panel_anterior = panel_anterior;
-    }
-    
-}
