+++ src/root/gestionagricola/vistas/supervisor/Busqueda.java
@@ -1,0 +1,152 @@
+
+package root.gestionagricola.vistas.supervisor;
+
+import javax.swing.JOptionPane;
+import root.gestionagricola.gestionsupervisor.ControladorAsistencia;
+import root.gestionagricola.vistas.ControladorVistas;
+
+/**
+ *
+ * @author Javier
+ */
+public class Busqueda extends javax.swing.JPanel {
+    
+    
+
+    private ControladorVistas controladorVista;
+    private String panel_anterior;
+    public Busqueda() {
+        initComponents();
+
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        Titulo7 = new javax.swing.JLabel();
+        RespuestaRutificador = new javax.swing.JTextField();
+        jButton4 = new javax.swing.JButton();
+        jButton6 = new javax.swing.JButton();
+        jLabel2 = new javax.swing.JLabel();
+        realizarBusqueda = new javax.swing.JButton();
+        Fondo = new javax.swing.JLabel();
+
+        setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
+
+        Titulo7.setBackground(new java.awt.Color(255, 255, 255));
+        Titulo7.setFont(new java.awt.Font("Garamond", 1, 36)); // NOI18N
+        Titulo7.setForeground(new java.awt.Color(255, 255, 255));
+        Titulo7.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        Titulo7.setText("Rutificador:");
+        add(Titulo7, new org.netbeans.lib.awtextra.AbsoluteConstraints(220, 290, 260, 60));
+        add(RespuestaRutificador, new org.netbeans.lib.awtextra.AbsoluteConstraints(510, 290, 330, 50));
+
+        jButton4.setIcon(new javax.swing.ImageIcon(getClass().getResource("/root/gestionagricola/vistas/imagenes/Volver atras.png"))); // NOI18N
+        jButton4.setBorderPainted(false);
+        jButton4.setContentAreaFilled(false);
+        jButton4.setFocusPainted(false);
+        jButton4.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton4ActionPerformed(evt);
+            }
+        });
+        add(jButton4, new org.netbeans.lib.awtextra.AbsoluteConstraints(1020, 30, 130, 70));
+
+        jButton6.setBackground(new java.awt.Color(255, 255, 255));
+        jButton6.setFont(new java.awt.Font("Garamond", 1, 36)); // NOI18N
+        jButton6.setIcon(new javax.swing.ImageIcon(getClass().getResource("/root/gestionagricola/vistas/imagenes/Busqueda 32px.png"))); // NOI18N
+        jButton6.setText("Búsqueda");
+        jButton6.setBorder(new javax.swing.border.LineBorder(new java.awt.Color(255, 153, 153), 2, true));
+        jButton6.setFocusPainted(false);
+        jButton6.setHorizontalTextPosition(javax.swing.SwingConstants.LEFT);
+        jButton6.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton6ActionPerformed(evt);
+            }
+        });
+        add(jButton6, new org.netbeans.lib.awtextra.AbsoluteConstraints(130, 110, 1020, 50));
+
+        jLabel2.setIcon(new javax.swing.ImageIcon(getClass().getResource("/root/gestionagricola/vistas/imagenes/frutap.png"))); // NOI18N
+        jLabel2.setText("jLabel2");
+        add(jLabel2, new org.netbeans.lib.awtextra.AbsoluteConstraints(-90, 0, 229, 151));
+
+        realizarBusqueda.setBackground(new java.awt.Color(255, 255, 255));
+        realizarBusqueda.setFont(new java.awt.Font("Garamond", 1, 24)); // NOI18N
+        realizarBusqueda.setText("Realizar Búsqueda");
+        realizarBusqueda.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                realizarBusquedaActionPerformed(evt);
+            }
+        });
+        add(realizarBusqueda, new org.netbeans.lib.awtextra.AbsoluteConstraints(460, 460, 260, 60));
+
+        Fondo.setIcon(new javax.swing.ImageIcon(getClass().getResource("/root/gestionagricola/vistas/imagenes/Fondo oscuro.png"))); // NOI18N
+        Fondo.setPreferredSize(new java.awt.Dimension(1299, 735));
+        add(Fondo, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 1270, 735));
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void jButton6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton6ActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jButton6ActionPerformed
+
+    private void jButton4ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton4ActionPerformed
+       //pueden acceder desde supervisor o dueño
+       //volver al dueño
+       if(this.panel_anterior.equals("busquedadueño")){
+           this.controladorVista.SeleccionarPanel("dueño");
+       }else if(this.panel_anterior.equals("busquedasupervisor")){//volver al supervisor
+            this.controladorVista.SeleccionarPanel("supervisor");
+       }
+       
+       
+
+    }//GEN-LAST:event_jButton4ActionPerformed
+
+    private void realizarBusquedaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_realizarBusquedaActionPerformed
+        
+        
+        if(this.RespuestaRutificador.getText() != null){
+            String[][] busqueda = ControladorAsistencia.realizarBusqueda(this.RespuestaRutificador.getText());
+            if(busqueda != null){
+                controladorVista.getAsistencia().reiniciarTabla();
+                controladorVista.getAsistencia().setDatos_tabla(busqueda);
+                controladorVista.getAsistencia().cargarDatosTabla();
+                controladorVista.SeleccionarPanel("asistencia");
+            }
+            else{
+                JOptionPane.showMessageDialog(null, " el rut ingresado no es válido", "Búsqueda", JOptionPane.INFORMATION_MESSAGE);
+            }
+        }
+        else{
+            JOptionPane.showMessageDialog(null, "Faltan llenar casillas.", "Búsqueda", JOptionPane.WARNING_MESSAGE);
+        }
+        
+        
+    }//GEN-LAST:event_realizarBusquedaActionPerformed
+
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JLabel Fondo;
+    private javax.swing.JTextField RespuestaRutificador;
+    private javax.swing.JLabel Titulo7;
+    private javax.swing.JButton jButton4;
+    private javax.swing.JButton jButton6;
+    private javax.swing.JLabel jLabel2;
+    private javax.swing.JButton realizarBusqueda;
+    // End of variables declaration//GEN-END:variables
+
+    public void setControladorVista(ControladorVistas controladorVista) {
+        this.controladorVista = controladorVista;
+    }
+
+    public void setPanel_anterior(String panel_anterior) {
+        this.panel_anterior = panel_anterior;
+    }
+    
+}
